+++ R/vi.R
@@ -78,6 +78,8 @@
 #'
 #' # Compute variable importance scores
 #' vi(mtcars.ppr, method = "ice")
+#' vi(mtcars.ppr, method = "ice",
+#'    var_fun = list("con" = mad, "cat" = function(x) diff(range(x)) / 4))
 #'
 #' # Plot variable importance scores
 #' vip(mtcars.ppr, method = "ice")